fix(zod): Type 'typeof dca' is not assignable to type 'BotTemplate<any>
