+++ packages/bot-templates/src/templates/dca.ts
@@ -55,47 +55,28 @@ dca.displayName = "DCA Bot";
 dca.description =
   "Dollar-Cost Averaging (DCA) is a trading strategy that involves entering a position through multiple smaller orders, known as Safety Orders. These orders are placed at predetermined levels, below the initial entry price. This method helps reduce the impact of adverse price movements by lowering the overall cost of the position. Once the market reverses and the price reaches a favorable level, the position is closed at the Take Profit level. This strategy is especially effective in volatile markets, allowing traders to capitalize on price fluctuations while minimizing the risks of poor timing with a single large entry.";
 dca.hidden = true;
-dca.schema = z
-  .object({
-    entry: z.object({
-      quantity: z.number().positive().describe("Quantity of the Entry Order in base currency"),
-      type: z.nativeEnum(XOrderType).describe("Entry with Limit or Market order"),
-      price: z.number().optional(),
-      conditions: z.any().optional(), // @todo schema validation
+dca.schema = z.object({
+  entry: z.object({
+    quantity: z.number().positive().describe("Quantity of the Entry Order in base currency"),
+    type: z.nativeEnum(XOrderType).describe("Entry with Limit or Market order"),
+    price: z.number().optional(),
+    conditions: z.any().optional(), // @todo schema validation
+  }),
+  tp: z.object({
+    percent: z.number().positive().describe("Take Profit from entry order price in %"),
+  }),
+  sl: z
+    .object({
+      percent: z.number().positive().describe("Stop Loss drop from entry order price in %"),
+    })
+    .optional(),
+  safetyOrders: z.array(
+    z.object({
+      quantity: z.number().positive().positive("Quantity of the Safety Order in base currency"),
+      priceDeviation: z.number().positive().positive("Price deviation from the Entry Order price in %"),
     }),
-    tp: z.object({
-      percent: z.number().positive().describe("Take Profit from entry order price in %"),
-    }),
-    sl: z
-      .object({
-        percent: z.number().positive().describe("Stop Loss drop from entry order price in %"),
-      })
-      .optional(),
-    safetyOrders: z.array(
-      z.object({
-        quantity: z.number().positive().positive("Quantity of the Safety Order in base currency"),
-        priceDeviation: z.number().positive().positive("Price deviation from the Entry Order price in %"),
-      }),
-    ),
-  })
-  .refine(
-    ({ sl, safetyOrders }) => {
-      // Validate SL: Ensure it is placed below the lowest Safety Order
-      if (sl && safetyOrders.length > 0) {
-        const lowestSafetyOrder = safetyOrders.reduce((acc, curr) => {
-          return curr.priceDeviation > acc.priceDeviation ? curr : acc;
-        });
-
-        // Note: `percent` and `priceDeviation` are positive numbers
-        if (sl.percent <= lowestSafetyOrder.priceDeviation) {
-          return false;
-        }
-      }
-
-      return true;
-    },
-    { message: `SL must be placed below the Safety Orders.` },
-  );
+  ),
+});
 
 dca.runPolicy = {
   onOrderFilled: true,
